requirements.txt
asgiref==3.4.1
black==21.8b0
Cerberus==1.3.4
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.7
django-admin-views==0.8.0
django-cors-headers==3.8.0
django-filter==21.1
django-heroku==0.3.1
django-multiselectfield==0.1.12
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.2
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==8.3.2
platformdirs==2.3.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
Pygments==2.10.0
PySocks==1.7.1
python-dotenv==0.19.0
pytz==2021.1
regex==2021.8.28
requests==2.26.0
requests-toolbelt==0.9.1
sqlparse==0.4.1
tomli==1.2.1
typing-extensions==3.10.0.2
urllib3==1.26.7
whitenoise==5.3.0
